analyze() and write_analysis()
FileWriter now has several overloaded functions called write_analysis() that can use any analyze() from the derived FileReader questions.
Still need to clean up comments and other surrounding code before merging branches. Also add tests.
