+++ FileManipulator.h
@@ -44,19 +44,6 @@ public:
 			Can we report how it failed (stderr) and why it returns false?
 	*/
 	virtual bool set_file_path(string newPath) = 0;
-	/**
-		Static function that determines if a path exists and can be opened for reading.
-		Parameters:
-			string newPath - Path to the file being read
-		Returns:
-			true - if the file can be opened and read
-			false - if the file does not exist, or it cannot be opened
-		Bugs:
-			none
-		Other:
-			Can we report how it failed (stderr) and why it returns false?
-	*/
-	static inline bool isValidFile(string newPath);
 
 	// Proposed Functions
 	/*
@@ -66,6 +53,20 @@ public:
 	static bool del_file();
 	static bool change_perms(); // file or "directory"
 	*/
+
+	/**
+	Static function that determines if a path exists and can be opened for reading.
+	Parameters:
+	string newPath - Path to the file being read
+	Returns:
+	true - if the file can be opened and read
+	false - if the file does not exist, or it cannot be opened
+	Bugs:
+	none
+	Other:
+	Can we report how it failed (stderr) and why it returns false?
+	*/
+	static bool isValidFile(string newPath);
 
 protected:
