+++ Source.cpp
@@ -16,28 +16,31 @@
 #include "FileWriter.h"
 #include "Test.h"
 
+#include "HydrogenBondReader.h"
+#include "SASAReader.h"
+#include "PDBReader.h"
+
 using namespace std;
 
 int main(int argc, char* argv[])
 {
 	try
 	{	
-		/*
-		FileReader f("C:/Users/Kevin/Documents/Github/k_amber-reader/data/sull_oxp-r-1.pdb");
-		FileWriter f_pdbFix("C:/Users/Kevin/Documents/Github/k_amber-reader/out/sull_oxp-r-1_fixed.pdb");
+		HydrogenBondReader hbr("C:/Users/Kevin/Documents/Github/k_amber-reader/data/SULL_ALP-R4_Hbond.dat");
+		SASAReader sasar("C:/Users/Kevin/Documents/Github/k_amber-reader/data/SULL_surface_ALP-R4_ligand.dat");
+		PDBReader pdbr("C:/Users/Kevin/Documents/Github/k_amber-reader/data/sull_oxp-r-1.pdb");
 
-		f.autofix_pdb("ULV");
-		f_pdbFix.write_autofix_pdb(f, "ULV");
-		*/
-
-		//f_pdbFix.autofix_pdb();
-
-		Test::test_reader_methods();
+		FileWriter fw("C:/Users/Kevin/Documents/Github/k_amber-reader/out/newbranch_sasa.txt");
+		//fw.write_analysis(sasar);
+		fw.set_file_path("C:/Users/Kevin/Documents/Github/k_amber-reader/out/newbranch_hbond.txt");
+		fw.write_analysis(hbr, 0.05);
+		fw.set_file_path("C:/Users/Kevin/Documents/Github/k_amber-reader/out/newbranch_pdb.txt");
+		fw.write_analysis(pdbr, "ULL");
 		
 	}
 	catch (Exception_FileManipulator& e)
 	{
-		e.what();
+		cout << e.what();
 	}
 
 	return 0;
