+++ Test.h
@@ -16,6 +16,16 @@ public:
 
 	static void test_Compute_conversions();
 
+	// individual classes
+	static void test_Compute();
+	static void test_ExceptionFileManipulator();
+	static void test_FileManipulator();
+	static void test_FileReader();
+	static void test_FileWriter();
+	static void test_HydrogenBondReader();
+	static void test_PDBReader();
+	static void test_SASAReader();
+
 private:
 
 	Test();
